move exporting functionality to core.model_export
