+++ resotolib/requirements.txt
@@ -5,7 +5,6 @@ psutil==5.9.0
 requests==2.27.1
 prometheus-client==0.13.1
 PyJWT==2.3.0
-typing-compat==0.1.0
 CherryPy==18.6.1
 cryptography==36.0.1
